Revert to previous version
