+++ app/src/main/java/me/devsaki/hentoid/activities/ASMHentaiActivity.java
@@ -1,10 +1,11 @@
-package me.devsaki.hentoid.activities.websites;
+package me.devsaki.hentoid.activities;
 
 import android.annotation.TargetApi;
 import android.graphics.Bitmap;
 import android.os.AsyncTask;
 import android.os.Build;
 import android.support.annotation.NonNull;
+import android.webkit.WebResourceError;
 import android.webkit.WebResourceRequest;
 import android.webkit.WebResourceResponse;
 import android.webkit.WebView;
@@ -14,8 +15,7 @@ import java.io.IOException;
 
 import me.devsaki.hentoid.database.domains.Content;
 import me.devsaki.hentoid.enums.Site;
-import me.devsaki.hentoid.parsers.HitomiParser;
-import me.devsaki.hentoid.util.ConstsPrefs;
+import me.devsaki.hentoid.parsers.ASMHentaiParser;
 import me.devsaki.hentoid.util.Helper;
 import me.devsaki.hentoid.views.ObservableWebView;
 import timber.log.Timber;
@@ -23,38 +23,24 @@ import timber.log.Timber;
 import static me.devsaki.hentoid.util.Helper.TYPE;
 import static me.devsaki.hentoid.util.Helper.executeAsyncTask;
 import static me.devsaki.hentoid.util.Helper.getWebResourceResponseFromAsset;
-import static me.devsaki.hentoid.util.Helper.getWebViewInitialZoomPrefs;
-import static me.devsaki.hentoid.util.Helper.getWebViewOverviewPrefs;
 
 /**
- * Created by Shiro on 1/20/2016.
- * Implements Hitomi.la source
+ * Created by avluis on 07/21/2016.
+ * Implements ASMHentai source
  */
-public class HitomiActivity extends BaseWebActivity {
+public class ASMHentaiActivity extends BaseWebActivity {
 
-    Site getStartSite() {
-        return Site.HITOMI;
+    @Override
+    void setSite(Site site) {
+        super.setSite(Site.ASMHENTAI);
     }
 
     @Override
     void setWebView(ObservableWebView webView) {
-        HitomiWebViewClient client = new HitomiWebViewClient();
-        client.restrictTo("hitomi.la");
+        ASMHentaiWebViewClient client = new ASMHentaiWebViewClient();
+        client.restrictTo("asmhentai.com");
 
         webView.setWebViewClient(client);
-
-        boolean bWebViewOverview = getWebViewOverviewPrefs();
-        int webViewInitialZoom = getWebViewInitialZoomPrefs();
-
-        if (bWebViewOverview) {
-            webView.getSettings().setLoadWithOverviewMode(false);
-            webView.setInitialScale(webViewInitialZoom);
-            Timber.d("WebView Initial Scale: %s%", webViewInitialZoom);
-        } else {
-            webView.setInitialScale(ConstsPrefs.PREF_WEBVIEW_INITIAL_ZOOM_DEFAULT);
-            webView.getSettings().setLoadWithOverviewMode(true);
-        }
-
         super.setWebView(webView);
     }
 
@@ -65,14 +51,21 @@ public class HitomiActivity extends BaseWebActivity {
         executeAsyncTask(new HtmlLoader(), extra);
     }
 
-    private class HitomiWebViewClient extends CustomWebViewClient {
+    private class ASMHentaiWebViewClient extends CustomWebViewClient {
         final ByteArrayInputStream nothing = new ByteArrayInputStream("".getBytes());
+
+        @Override
+        public void onReceivedError(WebView view, WebResourceRequest request,
+                                    WebResourceError error) {
+            /*Workaround for cache miss when re-submitting data to search form*/
+            view.loadUrl(view.getOriginalUrl());
+        }
 
         @Override
         public void onPageStarted(WebView view, String url, Bitmap favicon) {
             super.onPageStarted(view, url, favicon);
 
-            if (url.contains("//hitomi.la/galleries/")) {
+            if (url.contains("//asmhentai.com/g/") || url.contains("//comics.asmhentai.com/g/")) {
                 executeAsyncTask(new HtmlLoader(), url);
             }
         }
@@ -81,9 +74,12 @@ public class HitomiActivity extends BaseWebActivity {
         @Override
         public WebResourceResponse shouldInterceptRequest(@NonNull WebView view,
                                                           @NonNull String url) {
-            if (url.contains("hitomi.js")) {
-                return getWebResourceResponseFromAsset(getStartSite(), "hitomi.js", TYPE.JS);
-            } else if (url.contains("hitomi-horizontal.js") || url.contains("hitomi-vertical.js")) {
+            if (url.contains("ads.js") || url.contains("f.js") || url.contains("pop.js") ||
+                    url.contains("ads.php") || url.contains("syndication.exoclick.com")) {
+                return new WebResourceResponse("text/plain", "utf-8", nothing);
+            } else if (url.contains("main.js")) {
+                return getWebResourceResponseFromAsset(getSite(), "main.js", TYPE.JS);
+            } else if (url.contains("exoclick.com") || url.contains("juicyadultads.com")|| url.contains("exosrv.com")|| url.contains("hentaigold.net")) {
                 return new WebResourceResponse("text/plain", "utf-8", nothing);
             } else {
                 return super.shouldInterceptRequest(view, url);
@@ -95,9 +91,12 @@ public class HitomiActivity extends BaseWebActivity {
         public WebResourceResponse shouldInterceptRequest(@NonNull WebView view,
                                                           @NonNull WebResourceRequest request) {
             String url = request.getUrl().toString();
-            if (url.contains("hitomi.js")) {
-                return getWebResourceResponseFromAsset(getStartSite(), "hitomi.js", TYPE.JS);
-            } else if (url.contains("hitomi-horizontal.js") || url.contains("hitomi-vertical.js")) {
+            if (url.contains("ads.js") || url.contains("f.js") || url.contains("pop.js") ||
+                    url.contains("syndication.exoclick.com")) {
+                return new WebResourceResponse("text/plain", "utf-8", nothing);
+            } else if (url.contains("main.js")) {
+                return getWebResourceResponseFromAsset(getSite(), "main.js", TYPE.JS);
+            } else if (url.contains("exoclick.com") || url.contains("juicyadultads.com")|| url.contains("exosrv.com")|| url.contains("hentaigold.net")) {
                 return new WebResourceResponse("text/plain", "utf-8", nothing);
             } else {
                 return super.shouldInterceptRequest(view, request);
@@ -110,7 +109,7 @@ public class HitomiActivity extends BaseWebActivity {
         protected Content doInBackground(String... params) {
             String url = params[0];
             try {
-                processContent(HitomiParser.parseContent(url));
+                processContent(ASMHentaiParser.parseContent(url));
             } catch (IOException e) {
                 Timber.e(e, "Error parsing content.");
             }
