+++ app/src/main/java/me/devsaki/hentoid/activities/BaseWebActivity.java
@@ -1,4 +1,4 @@
-package me.devsaki.hentoid.activities.websites;
+package me.devsaki.hentoid.activities;
 
 import android.annotation.SuppressLint;
 import android.annotation.TargetApi;
@@ -25,7 +25,6 @@ import me.devsaki.hentoid.BuildConfig;
 import me.devsaki.hentoid.HentoidApp;
 import me.devsaki.hentoid.R;
 import me.devsaki.hentoid.abstracts.BaseActivity;
-import me.devsaki.hentoid.activities.DownloadsActivity;
 import me.devsaki.hentoid.database.HentoidDB;
 import me.devsaki.hentoid.database.domains.Content;
 import me.devsaki.hentoid.enums.Site;
@@ -49,15 +48,24 @@ import timber.log.Timber;
  * No particular source should be filtered/defined here.
  * The source itself should contain every method it needs to function.
  */
-public abstract class BaseWebActivity extends BaseActivity {
+public class BaseWebActivity extends BaseActivity {
 
     private Content currentContent;
     private HentoidDB db;
     private ObservableWebView webView;
+    private Site site;
     private boolean webViewIsLoading;
     private FloatingActionButton fabRead, fabDownload, fabRefreshOrStop, fabHome;
     private boolean fabReadEnabled, fabDownloadEnabled;
     private SwipeRefreshLayout swipeLayout;
+
+    Site getSite() {
+        return site;
+    }
+
+    void setSite(Site site) {
+        this.site = site;
+    }
 
     ObservableWebView getWebView() {
         return webView;
@@ -67,8 +75,6 @@ public abstract class BaseWebActivity extends BaseActivity {
         this.webView = webView;
     }
 
-    abstract Site getStartSite();
-
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -77,10 +83,11 @@ public abstract class BaseWebActivity extends BaseActivity {
 
         db = HentoidDB.getInstance(this);
 
-        if (getStartSite() == null) {
+        setSite(getSite());
+        if (site == null) {
             Timber.w("Site is null!");
         } else {
-            Timber.d("Loading site: %s", getStartSite());
+            Timber.d("Loading site: %s", site);
         }
 
         fabRead = (FloatingActionButton) findViewById(R.id.fabRead);
@@ -97,7 +104,7 @@ public abstract class BaseWebActivity extends BaseActivity {
         setWebView(getWebView());
 
         String intentVar = getIntent().getStringExtra(Consts.INTENT_URL);
-        webView.loadUrl(intentVar == null ? getStartSite().getUrl() : intentVar);
+        webView.loadUrl(intentVar == null ? site.getUrl() : intentVar);
     }
 
     @Override
@@ -135,7 +142,7 @@ public abstract class BaseWebActivity extends BaseActivity {
         webView.setOnLongClickListener(v -> {
             WebView.HitTestResult result = webView.getHitTestResult();
             if (result.getType() == WebView.HitTestResult.SRC_ANCHOR_TYPE) {
-                if (result.getExtra().contains(getStartSite().getUrl())) {
+                if (result.getExtra().contains(site.getUrl())) {
                     backgroundRequest(result.getExtra());
                 }
             } else {
@@ -366,7 +373,6 @@ public abstract class BaseWebActivity extends BaseActivity {
                 TsuminoParser.parseImageList(content);
                 break;
             case ASMHENTAI:
-            case ASMHENTAI_COMICS:
                 ASMHentaiParser.parseImageList(content);
                 break;
             case HENTAICAFE:
