+++ app/src/main/java/me/devsaki/hentoid/activities/HitomiActivity.java
@@ -1,4 +1,4 @@
-package me.devsaki.hentoid.activities.websites;
+package me.devsaki.hentoid.activities;
 
 import android.annotation.TargetApi;
 import android.graphics.Bitmap;
@@ -10,48 +10,71 @@ import android.webkit.WebResourceResponse;
 import android.webkit.WebView;
 
 import java.io.ByteArrayInputStream;
+import java.io.IOException;
 
 import me.devsaki.hentoid.database.domains.Content;
 import me.devsaki.hentoid.enums.Site;
-import me.devsaki.hentoid.parsers.NhentaiParser;
-import me.devsaki.hentoid.util.HttpClientHelper;
+import me.devsaki.hentoid.parsers.HitomiParser;
+import me.devsaki.hentoid.util.ConstsPrefs;
+import me.devsaki.hentoid.util.Helper;
 import me.devsaki.hentoid.views.ObservableWebView;
 import timber.log.Timber;
 
 import static me.devsaki.hentoid.util.Helper.TYPE;
 import static me.devsaki.hentoid.util.Helper.executeAsyncTask;
 import static me.devsaki.hentoid.util.Helper.getWebResourceResponseFromAsset;
+import static me.devsaki.hentoid.util.Helper.getWebViewInitialZoomPrefs;
+import static me.devsaki.hentoid.util.Helper.getWebViewOverviewPrefs;
 
 /**
  * Created by Shiro on 1/20/2016.
- * Implements nhentai source
+ * Implements Hitomi.la source
  */
-public class NhentaiActivity extends BaseWebActivity {
+public class HitomiActivity extends BaseWebActivity {
 
-    Site getStartSite() {
-        return Site.NHENTAI;
+    @Override
+    void setSite(Site site) {
+        super.setSite(Site.HITOMI);
     }
 
     @Override
     void setWebView(ObservableWebView webView) {
-        NhentaiWebViewClient client = new NhentaiWebViewClient();
-        client.restrictTo("nhentai.net");
+        HitomiWebViewClient client = new HitomiWebViewClient();
+        client.restrictTo("hitomi.la");
 
         webView.setWebViewClient(client);
+
+        boolean bWebViewOverview = getWebViewOverviewPrefs();
+        int webViewInitialZoom = getWebViewInitialZoomPrefs();
+
+        if (bWebViewOverview) {
+            webView.getSettings().setLoadWithOverviewMode(false);
+            webView.setInitialScale(webViewInitialZoom);
+            Timber.d("WebView Initial Scale: %s%", webViewInitialZoom);
+        } else {
+            webView.setInitialScale(ConstsPrefs.PREF_WEBVIEW_INITIAL_ZOOM_DEFAULT);
+            webView.getSettings().setLoadWithOverviewMode(true);
+        }
+
         super.setWebView(webView);
     }
 
-    private class NhentaiWebViewClient extends CustomWebViewClient {
+    @Override
+    void backgroundRequest(String extra) {
+        Timber.d(extra);
+        Helper.toast("Processing...");
+        executeAsyncTask(new HtmlLoader(), extra);
+    }
+
+    private class HitomiWebViewClient extends CustomWebViewClient {
         final ByteArrayInputStream nothing = new ByteArrayInputStream("".getBytes());
 
         @Override
         public void onPageStarted(WebView view, String url, Bitmap favicon) {
             super.onPageStarted(view, url, favicon);
 
-            if (url.contains("nhentai.net/g/")) {
-                String newURL = url.replace("/g", "/api/gallery");
-                newURL = newURL.substring(0, newURL.length() - 1);
-                executeAsyncTask(new JsonLoader(), newURL);
+            if (url.contains("//hitomi.la/galleries/")) {
+                executeAsyncTask(new HtmlLoader(), url);
             }
         }
 
@@ -59,11 +82,9 @@ public class NhentaiActivity extends BaseWebActivity {
         @Override
         public WebResourceResponse shouldInterceptRequest(@NonNull WebView view,
                                                           @NonNull String url) {
-            if (url.contains("//static.nhentai.net/js/")) {
-                return getWebResourceResponseFromAsset(getStartSite(), "main_js.js", TYPE.JS);
-            } else if (url.contains("//static.nhentai.net/css/")) {
-                return getWebResourceResponseFromAsset(getStartSite(), "main_style.css", TYPE.CSS);
-            } else if (url.contains("ads.contentabc.com")) {
+            if (url.contains("hitomi.js")) {
+                return getWebResourceResponseFromAsset(getSite(), "hitomi.js", TYPE.JS);
+            } else if (url.contains("hitomi-horizontal.js") || url.contains("hitomi-vertical.js")) {
                 return new WebResourceResponse("text/plain", "utf-8", nothing);
             } else {
                 return super.shouldInterceptRequest(view, url);
@@ -75,11 +96,9 @@ public class NhentaiActivity extends BaseWebActivity {
         public WebResourceResponse shouldInterceptRequest(@NonNull WebView view,
                                                           @NonNull WebResourceRequest request) {
             String url = request.getUrl().toString();
-            if (url.contains("//static.nhentai.net/js/")) {
-                return getWebResourceResponseFromAsset(getStartSite(), "main_js.js", TYPE.JS);
-            } else if (url.contains("//static.nhentai.net/css/")) {
-                return getWebResourceResponseFromAsset(getStartSite(), "main_style.css", TYPE.CSS);
-            } else if (url.contains("ads2.contentabc.com")) {
+            if (url.contains("hitomi.js")) {
+                return getWebResourceResponseFromAsset(getSite(), "hitomi.js", TYPE.JS);
+            } else if (url.contains("hitomi-horizontal.js") || url.contains("hitomi-vertical.js")) {
                 return new WebResourceResponse("text/plain", "utf-8", nothing);
             } else {
                 return super.shouldInterceptRequest(view, request);
@@ -87,13 +106,13 @@ public class NhentaiActivity extends BaseWebActivity {
         }
     }
 
-    private class JsonLoader extends AsyncTask<String, Integer, Content> {
+    private class HtmlLoader extends AsyncTask<String, Integer, Content> {
         @Override
         protected Content doInBackground(String... params) {
             String url = params[0];
             try {
-                processContent(NhentaiParser.parseContent(HttpClientHelper.call(url)));
-            } catch (Exception e) {
+                processContent(HitomiParser.parseContent(url));
+            } catch (IOException e) {
                 Timber.e(e, "Error parsing content.");
             }
