+++ app/src/main/java/me/devsaki/hentoid/parsers/ASMHentaiParser.java
@@ -1,7 +1,5 @@
 package me.devsaki.hentoid.parsers;
 
-import android.webkit.URLUtil;
-
 import org.jsoup.Jsoup;
 import org.jsoup.nodes.Document;
 import org.jsoup.nodes.Element;
@@ -14,7 +12,6 @@ import java.util.List;
 import me.devsaki.hentoid.database.domains.Attribute;
 import me.devsaki.hentoid.database.domains.Content;
 import me.devsaki.hentoid.enums.AttributeType;
-import me.devsaki.hentoid.enums.Site;
 import me.devsaki.hentoid.enums.StatusContent;
 import me.devsaki.hentoid.util.AttributeMap;
 import timber.log.Timber;
@@ -27,7 +24,7 @@ import static me.devsaki.hentoid.enums.Site.ASMHENTAI;
  */
 public class ASMHentaiParser {
 
-    public static Content parseContent(String urlString, Site site) throws IOException {
+    public static Content parseContent(String urlString) throws IOException {
         Document doc = Jsoup.connect(urlString).get();
 
         Elements content = doc.select("div.info");
@@ -92,7 +89,7 @@ public class ASMHentaiParser {
                     .setAttributes(attributes)
                     .setQtyPages(pages)
                     .setStatus(StatusContent.SAVED)
-                    .setSite(site);
+                    .setSite(ASMHENTAI);
         }
 
         return null;
@@ -119,12 +116,6 @@ public class ASMHentaiParser {
         int pages = content.getQtyPages();
         String readerUrl = content.getReaderUrl();
         List<String> imgUrls = new ArrayList<>();
-
-        if (!URLUtil.isValidUrl(readerUrl))
-        {
-            Timber.e("Invalid URL : %s", readerUrl);
-            return imgUrls;
-        }
 
         Document doc;
         String ext;
