+++ app/src/main/res/xml/app_tracker.xml
@@ -20,13 +20,13 @@
     <screenName name="me.devsaki.hentoid.activities.DownloadsActivity">Home: Downloads</screenName>
     <screenName name="me.devsaki.hentoid.activities.QueueActivity">Home: Queue</screenName>
 
-    <screenName name="me.devsaki.hentoid.activities.websites.BaseWebActivity">Source: Legacy</screenName>
-    <screenName name="me.devsaki.hentoid.activities.websites.HitomiActivity">Source: hitomi</screenName>
-    <screenName name="me.devsaki.hentoid.activities.websites.NhentaiActivity">Source: nhentai</screenName>
-    <screenName name="me.devsaki.hentoid.activities.websites.TsuminoActivity">Source: tsumino</screenName>
-    <screenName name="me.devsaki.hentoid.activities.websites.ASMHentaiActivity">Source: ASMHentai</screenName>
-    <screenName name="me.devsaki.hentoid.activities.websites.HentaiCafeActivity">Source: Hentai Cafe</screenName>
-    <screenName name="me.devsaki.hentoid.activities.websites.PururinActivity">Source: Pururin</screenName>
+    <screenName name="me.devsaki.hentoid.activities.BaseWebActivity">Source: Legacy</screenName>
+    <screenName name="me.devsaki.hentoid.activities.HitomiActivity">Source: hitomi</screenName>
+    <screenName name="me.devsaki.hentoid.activities.NhentaiActivity">Source: nhentai</screenName>
+    <screenName name="me.devsaki.hentoid.activities.TsuminoActivity">Source: tsumino</screenName>
+    <screenName name="me.devsaki.hentoid.activities.ASMHentaiActivity">Source: ASMHentai</screenName>
+    <screenName name="me.devsaki.hentoid.activities.HentaiCafeActivity">Source: Hentai Cafe</screenName>
+    <screenName name="me.devsaki.hentoid.activities.PururinActivity">Source: Pururin</screenName>
 
     <screenName name="me.devsaki.hentoid.activities.IntentActivity">Launch: Intent</screenName>
     <screenName name="me.devsaki.hentoid.activities.SplashActivity">Launch: Splash</screenName>
